feat(core): Add core components for game engine
This commit consists of the initial implementation of core components
for the game engine. In particular, a standard GameObject is created
which ensures the desginer can implemented multiple characteristics.

* To that, the Tile class is responsible for rendering a simple tile
  that can be colored. In the future, this will not be required as
theoretically everything can be based off GameObject alone. This class
simplifies testing the rendering of a single colorful tile.

Additionally, the MainEngine has been created which links the python
scripting language to the game engine. This is necessary as we would
like to use a scripting language to design the game. The current
language is python but can be changed to another high level language.

Util class implements helpful mem checks which is good practice to
ensure memory is being allocated properly. Checking free pointers though
not required, may help to determine if certain memory blocks are
properly freed.

Subsequently, the Test files have been added with simple unit tests to
gauge working functionality. Additionally, the Test directory has been
updated to be modularized and directly refers to the component being
tested.