feat(MemoryManagement): Add memory management integration
This commit adds in Memory Management handler which utilizes similar
"new" and "delete" but with mimalloc. Additionally, logic changes have
been improved in Main.cpp. I have noticed that SDL_RenderFillRect()
produces an side-effect as it tries to free the object passed. I was
able to figure this after using Valgrind which pointed to the direction
causing this. I strongly believe that unique_ptr should be used
sparingly but this would be fine as we build out Heap and Stack to draw
the levels.

The changes to the commit include:
* Creation of MemoryManagement.cpp
* Addition of DeepCopy template to Util
* Integration of MemoryManagement with GameObject and Tile Classes

The purpose of MemoryManagement is to provide a safe method for
allocations which prevents access to unwanted memory. Note, that this
works with the destructor which means there's no need to deallocate
anything manually. This is exactly what caused the double free as blocks
freed contain the already freed rect from SDL.