feat(GameAction), bug(Event): Fix event system
This commit fixes the previous by ensuring that the controller class now
works. Game objects are subscribed to the Handler, and the Controller
notifies the EventHandler when input is detected. The event system is an
important feature which allows modification to be done on an object,
generically, and can be customized depending on needs. Given this, this
commit (and previous) are core features to the game engine.

Important changes:
MemoryManagement and Map used to assign the responsibility of the
objects to the classes themselves. This is a problem when using the
data structure for references only. To accomplish this, the parameter
"CC" or Complete Clean, has been added and is default set to true for
both classes. This parameter is responsibile for clearing subobjects
within memory. Disabling this prevents this from happening, but that
means the responsibility of the object is *delagated* elsewhere.