feat(animation): Add animation, bug(controller)
This commit adds in animation and cels related to the rendering. This
interacts with the rendering engine by sending over the cels depending
on the state observed by the handler.

Additionally, this commit addresses a bug with the Controller class
where setting an EventHandler can cause memory leaks.

Lastly, all test files have been included and updated to the latest
configurations.